requirements.txt
attrs==21.4.0
awkward==1.8.0
bleach==4.1.0
build==0.7.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
commonmark==0.9.1
contextlib2==21.6.0
cryptography==36.0.2
cycler==0.11.0
datenraffinerie @ file:///home/nabla/cern/datenraffinerie
docutils==0.18.1
fonttools==4.33.3
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
jeepney==0.8.0
Jinja2==3.1.1
keyring==23.5.0
kiwisolver==1.4.2
llvmlite==0.38.0
lockfile==0.12.2
luigi==3.0.3
MarkupSafe==2.1.1
matplotlib==3.5.1
nested-dict==1.61
numba==0.55.1
numexpr==2.8.1
numpy==1.21.6
packaging==21.3
pandas==1.4.2
pep517==0.12.0
Pillow==9.1.0
pkginfo==1.8.2
pluggy==1.0.0
py==1.11.0
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.7
pytest==7.1.2
python-daemon==2.3.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
readme-renderer==34.0
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.2.0
schema==0.7.5
scipy==1.8.0
SecretStorage==3.3.1
six==1.16.0
tables==3.7.0
tenacity==6.3.1
tomli==2.0.1
tornado==6.1
twine==4.0.0
uproot==4.2.2
urllib3==1.26.9
webencodings==0.5.1
zipp==3.8.0
zmq==0.0.0
